Module Tezos_dac_client_lib.Dac_node_client

class type cctxt = object ... end

Instance of Tezos_client_base.Client_context that only handles IOs and RPCs. Can be used for keys and RPCs related commands.

class unix_cctxt : rpc_config:Tezos_rpc_http_client_unix.RPC_client_unix.config -> cctxt

Instance of cctxt for linux systems. Relies on Tezos_rpc_http_client_unix.

val make_unix_cctxt : scheme:string -> host:string -> port:int -> cctxt

make_unix_client_context scheme host port generates a cctxt from the client configuration parameters.

val of_uri : Uri.t -> cctxt

of_uri uri generates a cctxt from uri.

module V0 : sig ... end

V0 is a module that provides a client specification for interacting with experimental Tezos_lib_dac.Rpc_services.Api.V0 API. Note that even though V0 api is binding, it is already deprecated. Use it at your own risk!