Module Dac_node_client.V0

V0 is a module that provides a client specification for interacting with experimental Tezos_lib_dac.Rpc_services.Api.V0 API. Note that even though V0 api is binding, it is already deprecated. Use it at your own risk!

val get_preimage : cctxt -> page_hash:Tezos_dac_lib.Dac_plugin.raw_hash -> bytes Tezos_base.TzPervasives.tzresult Lwt.t

get_preimage cctxt ~hash requests the preimage of hash, consisting of a single page, from cctxt. When the request succeeds, the raw page will be returned as a sequence of bytes.

val put_dac_member_signature : cctxt -> signature:Tezos_dac_lib.Signature_repr.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

put_dac_member_signature cctxt ~signature:Signature_repr.t stores the signature generated from signing hex_root_hash by dac_member_pkh.

val get_certificate : cctxt -> root_page_hash:Tezos_dac_lib.Dac_plugin.raw_hash -> Tezos_dac_lib.Certificate_repr.t Tezos_base.TzPervasives.tzresult Lwt.t

get_certificate cctxt ~root_page_hash fetches the DAC certificate for the provided root_page_hash.

val get_serialized_certificate : cctxt -> root_page_hash:Tezos_dac_lib.Dac_plugin.raw_hash -> Tezos_base.TzPervasives.String.t option Tezos_base.TzPervasives.tzresult Lwt.t

get_serialized_certificate cctxt ~root_page_hash fetches and serialize the DAC certificate for the provided root_page_hash.

val monitor_certificate : cctxt -> root_hash:Tezos_dac_lib.Dac_plugin.raw_hash -> (Tezos_dac_lib.Certificate_repr.t Lwt_stream.t * Tezos_rpc.Context.stopper) Tezos_base.TzPervasives.tzresult Lwt.t

monitor_certificate cctxt ~root_hash returns a stream and a stopper for monitoring certificate updates for root_hash.

module Coordinator : sig ... end

module Observer : sig ... end