Module V0.Coordinator

val post_preimage : cctxt -> payload:bytes -> Tezos_dac_lib.Dac_plugin.raw_hash Tezos_base.TzPervasives.tzresult Lwt.t

post_preimage cctxt ~payload sends a payload to the DAC Coordinator via a POST RPC call to v0/preimage. It returns a hex encoded root page hash, produced by Merkle_tree_V0 pagination scheme. On the backend side it also pushes root page hash of the preimage to all the subscribed DAC Members and Observers.