Module V0.Observer

val get_missing_page : cctxt -> Tezos_dac_lib.Dac_plugin.raw_hash -> bytes Tezos_base.TzPervasives.tzresult Lwt.t

get_missing_page plugin cctxt page_hash notifies an Observer that a page hash must be downloaded. This function returns the preimage of the requested page_hash.